refactor(boardName)[OK-212] Split code into 2 files
